Remove dynamic security analysis Client
